requirements.txt
Flask
psycopg2-binary
tensorflow
opencv-python
numpy
Werkzeug
